add: 服に対応 already_asked削除処理追加

+++ alarm/alarm.py
@@ -14,6 +14,8 @@ def main():
     schedule.every().day.at(get_up_time).do(sound)
     # アラーム待ち
     while True:
+        if not os.path.isfile("../alarm/alarm_set.dat"):
+            return schedule.CancelJob()
         schedule.run_pending()
         time.sleep(1)
 
@@ -36,12 +38,13 @@ def get_time(string):
 def sound():
     while True:
         os.system("play ../alarm/alarm1.mp3")  # パス指定必要な場合はここで
+        if os.path.isfile("../alarm/alarm_set.dat"):
+            with open("../alarm/alarm_set.dat","w") as f:
+                alarm_ringed = str(1)
+                f.write(alarm_ringed)
         time.sleep(1)
-        with open("../alarm/alarm_set.dat", "w") as f:
-            alarm_ringed = str(1)
-            f.write(alarm_ringed)
         if not os.path.isfile("../alarm/alarm_set.dat"):
-            break
+            return schedule.CancelJob()
 
 
 if __name__ == "__main__":
